move files to aws-cdk-scala
