+++ modules/aws-cdk-scala/src/main/scala/io/burkard/cdk/package.scala
